unify invoicing_disabled_reports
