+++ leasing/report/utils.py
@@ -1,0 +1,5 @@
+# From Django docs
+def dictfetchall(cursor):
+    """Return all rows from a cursor as a dict"""
+    columns = [col[0] for col in cursor.description]
+    return [dict(zip(columns, row)) for row in cursor.fetchall()]
